+++ keba_keenergy_api/endpoints.py
@@ -31,6 +31,7 @@ from keba_keenergy_api.constants import HeatPumpCompressorUseNightSpeed
 from keba_keenergy_api.constants import HeatPumpOperatingMode
 from keba_keenergy_api.constants import HotWaterTank
 from keba_keenergy_api.constants import HotWaterTankOperatingMode
+from keba_keenergy_api.constants import HotWaterTankUseExcessEnergy
 from keba_keenergy_api.constants import LineTablePool
 from keba_keenergy_api.constants import MAX_HEATING_CURVE_POINTS
 from keba_keenergy_api.constants import MIN_HEATING_CURVE_POINTS
@@ -1659,6 +1660,88 @@ class HotWaterTankEndpoints(BaseEndpoints):
         """
         temperatures: list[float | None] = [temperature if position == p else None for p in range(1, position + 1)]
         await self._write_values(request={HotWaterTank.EXCESS_ENERGY_TARGET_TEMPERATURE_HYSTERESIS: temperatures})
+
+    async def get_use_excess_energy(
+        self,
+        position: int = 1,
+        *,
+        human_readable: bool = True,
+    ) -> int | str:
+        """Get the use excess energy state.
+
+        Parameters
+        ----------
+        position
+            The number of the hot water tanks
+        human_readable
+            Return a human-readable string
+
+        Returns
+        -------
+        integer or string
+            (0) OFF / (1) ON
+
+        """
+        response: dict[str, list[list[Value]] | list[Value]] = await self._read_data(
+            request=HotWaterTank.USE_EXCESS_ENERGY,
+            position=position,
+            human_readable=human_readable,
+            extra_attributes=True,
+        )
+        return self._get_int_or_str_value(response, section=HotWaterTank.USE_EXCESS_ENERGY, position=position)
+
+    async def set_use_excess_energy(self, mode: int | str, position: int = 1) -> None:
+        """Set the use excess energy.
+
+        **Attention!** Writing values should remain within normal limits, as is the case with typical use of the
+        Web HMI. Permanent and very frequent writing of values reduces the lifetime of the built-in flash memory.
+
+        Parameters
+        ----------
+        mode
+            Set the mode as integer or string (human-readable) e.g. 0 or OFF
+        position
+            The number of the hot water tanks
+
+        """
+        try:
+            _mode: int | None = mode if isinstance(mode, int) else HotWaterTankUseExcessEnergy[mode.upper()].value
+        except KeyError as error:
+            message: str = f"Invalid value! Allowed values are {self._get_allowed_values(HotWaterTankUseExcessEnergy)}"
+            raise APIError(message) from error
+
+        modes: list[int | None] = [_mode if position == p else None for p in range(1, position + 1)]
+
+        await self._write_values(request={HotWaterTank.USE_EXCESS_ENERGY: modes})
+
+    async def get_excess_energy_available(
+        self,
+        position: int = 1,
+        *,
+        human_readable: bool = True,
+    ) -> int | str:
+        """Get the excess energy available state.
+
+        Parameters
+        ----------
+        position
+            The number of the hot water tanks
+        human_readable
+            Return a human-readable string
+
+        Returns
+        -------
+        integer or string
+            (0) OFF / (1) ON
+
+        """
+        response: dict[str, list[list[Value]] | list[Value]] = await self._read_data(
+            request=HotWaterTank.EXCESS_ENERGY_AVAILABLE,
+            position=position,
+            human_readable=human_readable,
+            extra_attributes=True,
+        )
+        return self._get_int_or_str_value(response, section=HotWaterTank.EXCESS_ENERGY_AVAILABLE, position=position)
 
     async def get_heat_request(self, position: int = 1, *, human_readable: bool = True) -> int | str:
         """Get the heat request state from the hot water tank.