+++ CHANGELOG.md
@@ -27,6 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_excess_energy_target_temperature_hysteresis()` to the hot water tank
 - Added endpoint `get_excess_energy_target_temperature()` to the hot water tank
 - Added endpoint `set_excess_energy_target_temperature()` to the hot water tank
+- Added endpoint `get_excess_energy_available()` to the hot water tank
+- Added endpoint `get_use_excess_energy()` to the hot water tank
+- Added endpoint `set_use_excess_energy()` to the hot water tank
 - Added endpoint `get_excess_energy_target_temperature()` to the heating circuit
 - Added endpoint `set_excess_energy_target_temperature()` to the heating circuit
 - Added endpoint `get_excess_energy_target_temperature_hysteresis()` to the heating circuit